whatsapp_scheduler/requirements.txt
# Core PydanticAI framework
pydantic-ai>=0.0.14
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Model providers
openai>=1.0.0
anthropic>=0.20.0

# Environment and configuration
python-dotenv>=1.0.0

# HTTP client for API calls
httpx>=0.25.0

# Flask for webhook server
flask>=2.3.0

# Natural language date parsing
dateparser>=1.1.0

# Database (optional - SQLite is built-in, but for PostgreSQL)
# psycopg2-binary>=2.9.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development and debugging
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0